engine/requirements.txt
# Blood Ledger Backend Dependencies

# Core
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pyyaml>=6.0.0

# FalkorDB (direct Python client)
falkordb>=1.0.0

# Embeddings
sentence-transformers>=2.2.0
numpy>=1.24.0

# LLM (Claude CLI is used via subprocess)
# No Python dependency needed

# Utilities
python-dotenv>=1.0.0
